Change to new project
Swapping over to a completely new project. This was due to the fact that the past project was obsolete and incompatible with 2024 WPILIB, and wouldn't build.

Remapped motor IDs to corresponding IDs on the robot

Tuned P value in teleopswerve to counteract spinning in place while driving forward.

Added AutoConstants to swerveconfig file to make autonomous operational. Added exampleAutos file and introduced a simple S curve autonomous.